+++ crate_anon/nlprp/api.py
@@ -41,6 +41,7 @@ from cardinal_pythonlib.datetimefunc import (
     pendulum_to_datetime,
     pendulum_to_utc_datetime_without_tz,
 )
+from cardinal_pythonlib.httpconst import HttpStatus
 from cardinal_pythonlib.json.typing_helpers import (
     JsonArrayType,
     JsonAsStringType,
@@ -54,7 +55,6 @@ from semantic_version import Version
 
 from crate_anon.common.constants import JSON_SEPARATORS_COMPACT
 from crate_anon.nlprp.constants import (
-    HttpStatus,
     NlprpKeys,
     NlprpValues,
     ALL_NLPRP_COMMANDS,